backend/coffee_backend/requirements.txt
Django>=5.1.7
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
channels>=4.0.0
gunicorn>=23.0.0
djangorestframework-simplejwt>=5.2.2
pyotp>=2.8.0
qrcode>=7.4.2
Pillow>=10.0.0 